+++ apps/openchallenges/kaggle-to-kafka-service/src/main/java/org/sagebionetworks/openchallenges/kaggle/to/kafka/service/exception/KaggleToKafkaServiceException.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.openchallenges.exception;
+package org.sagebionetworks.openchallenges.kaggle.to.kafka.service.exception;
 
 public class KaggleToKafkaServiceException extends RuntimeException {
