+++ apps/openchallenges/kaggle-to-kafka-service/src/main/java/org/sagebionetworks/openchallenges/kaggle/to/kafka/service/initializer/StreamInitializer.java
@@ -1,0 +1,5 @@
+package org.sagebionetworks.openchallenges.kaggle.to.kafka.service.initializer;
+
+public interface StreamInitializer {
+  void init();
+}
