+++ apps/openchallenges/kaggle-to-kafka-service/src/main/java/org/sagebionetworks/openchallenges/kaggle/to/kafka/service/runner/StreamRunner.java
@@ -1,0 +1,5 @@
+package org.sagebionetworks.openchallenges.kaggle.to.kafka.service.runner;
+
+public interface StreamRunner {
+  void start();
+}
