--- apps/openchallenges/kaggle-to-kafka-service/src/main/java/org/sagebionetworks/openchallenges/runner/StreamRunner.java
@@ -1,5 +1,0 @@
-package org.sagebionetworks.openchallenges.runner;
-
-public interface StreamRunner {
-  void start();
-}
